requirements.txt
Flask==2.0.1
requests
gunicorn==20.1.0
python-dotenv
pymlask==0.3.2
mecab-python3==1.0.4
numpy==1.20.2
